Problem set 0
This file contains the code to the problem set 0 using scratch. It is archery game, where the player is given 10 attempts to shoot the target in order to win the game. 
The game starts when the player presses the green flag, then the arrow is launched when clicking the space key. 
When 10 attempts reached, the message "You won" is displayed when the players scores 10 successful shots, otherwise "Game Over" is displayed.